drone-ops-coordinator/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.25.0
httpx==0.27.0
pydantic>=2.0
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
pandas>=2.2.2
numpy>=1.26.0
pytest==7.4.3
requests==2.31.0